BasketResponse sınıfında basket ve customer id tipleri değitirildi ve ProductDto da yapıcı metot oluşturuldu

+++ src/main/java/org/example/eticaret/response/BasketResponse.java
@@ -11,9 +11,9 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BasketResponse {
-    private Long basketId;
+    private int basketId;
     private double totalPrice;
     private int status;
-    private Long customerId;
+    private int customerId;
     private List<BasketItemDto> basketItemDtoList;
 }
